+++ packages/gatsby-transformer-json/src/__tests__/gatsby-node.js
@@ -1,4 +1,5 @@
 const Promise = require(`bluebird`)
+const os = require(`os`)
 
 const { onCreateNode } = require(`../gatsby-node`)
 
@@ -11,6 +12,7 @@ const bootstrapTest = async (node, pluginOptions = {}) => {
   const actions = { createNode, createParentChildLink }
   const createNodeId = jest.fn()
   createNodeId.mockReturnValue(`uuid-from-gatsby`)
+  const createContentDigest = jest.fn().mockReturnValue(`contentDigest`)
 
   return await onCreateNode(
     {
@@ -18,6 +20,7 @@ const bootstrapTest = async (node, pluginOptions = {}) => {
       loadNodeContent,
       actions,
       createNodeId,
+      createContentDigest,
     },
     pluginOptions
   ).then(() => {
@@ -31,7 +34,7 @@ const bootstrapTest = async (node, pluginOptions = {}) => {
 describe(`Process JSON nodes correctly`, () => {
   const baseNode = {
     id: `whatever`,
-    parent: `SOURCE`,
+    parent: null,
     children: [],
     internal: {
       contentDigest: `whatever`,
@@ -42,7 +45,7 @@ describe(`Process JSON nodes correctly`, () => {
   const baseFileNode = {
     ...baseNode,
     name: `nodeName`,
-    dir: `/tmp/foo/`,
+    dir: `${os.tmpdir()}/foo/`,
     internal: {
       ...baseNode.internal,
       type: `File`,