+++ packages/gatsby/src/commands/repl.js
@@ -1,7 +1,13 @@
 const repl = require(`repl`)
 const { graphql } = require(`graphql`)
 const bootstrap = require(`../bootstrap`)
-const { store, loadNodeContent, getNodes, getNode } = require(`../redux`)
+const {
+  loadNodeContent,
+  getNodes,
+  getNode,
+  getNodesByType,
+} = require(`../db/nodes`)
+const { store } = require(`../redux`)
 
 module.exports = async program => {
   // run bootstrap
@@ -12,12 +18,12 @@ module.exports = async program => {
     schema,
     config,
     babelrc,
-    jsonDataPaths,
     pages,
     components,
     staticQueryComponents,
-    nodes,
   } = store.getState()
+
+  const nodes = getNodes()
 
   const query = async query => {
     const result = await graphql(schema, query, {}, {}, {})
@@ -32,9 +38,9 @@ module.exports = async program => {
   // set some globals to make life easier
   _.context.babelrc = babelrc
   _.context.components = components
-  _.context.dataPaths = jsonDataPaths
   _.context.getNode = getNode
   _.context.getNodes = getNodes
+  _.context.getNodesByType = getNodesByType
   _.context.loadNodeContent = loadNodeContent
   _.context.nodes = [...nodes.entries()]
   _.context.pages = [...pages.entries()]