+++ www/src/components/layout.js
@@ -1,25 +1,34 @@
 import React from "react"
-import Modal from "react-modal"
-import Helmet from "react-helmet"
-import MdClose from "react-icons/lib/md/close"
 import { navigate, PageRenderer } from "gatsby"
-import presets, { colors } from "../utils/presets"
+import mousetrap from "mousetrap"
+import Modal from "react-modal"
+import { SkipNavLink } from "@reach/skip-nav"
+import MdClose from "react-icons/lib/md/close"
+
+import {
+  colors,
+  radii,
+  space,
+  shadows,
+  sizes,
+  fontSizes,
+  zIndices,
+} from "../utils/presets"
+import { breakpointGutter } from "../utils/styles"
 import Banner from "../components/banner"
 import Navigation from "../components/navigation"
 import MobileNavigation from "../components/navigation-mobile"
 import PageWithSidebar from "../components/page-with-sidebar"
-
-import mousetrap from "mousetrap"
+import SiteMetadata from "../components/site-metadata"
 
 // Import Futura PT typeface
 import "../fonts/Webfonts/futurapt_book_macroman/stylesheet.css"
 import "../fonts/Webfonts/futurapt_bookitalic_macroman/stylesheet.css"
 import "../fonts/Webfonts/futurapt_demi_macroman/stylesheet.css"
 import "../fonts/Webfonts/futurapt_demiitalic_macroman/stylesheet.css"
+import "../fonts/Webfonts/futurapt_bold_macroman/stylesheet.css"
 
-// Other fonts
-import "typeface-spectral"
-import "typeface-space-mono"
+import { skipLink } from "../utils/styles"
 
 let windowWidth
 
@@ -56,8 +65,6 @@ class DefaultLayout extends React.Component {
   }
 
   render() {
-    const isHomepage = this.props.location.pathname === `/`
-
     // SEE: template-docs-markdown for why this.props.isSidebarDisabled is here
     const isSidebarDisabled =
       this.props.isSidebarDisabled || !this.props.itemList
@@ -71,7 +78,7 @@ class DefaultLayout extends React.Component {
 
     if (isModal && window.innerWidth > 750) {
       return (
-        <React.Fragment>
+        <>
           <PageRenderer
             location={{ pathname: this.props.modalBackgroundPath }}
           />
@@ -87,7 +94,7 @@ class DefaultLayout extends React.Component {
                 width: `750px`,
                 background: `none`,
                 border: `none`,
-                padding: `40px 0`,
+                padding: `${space[8]} 0`,
                 overflow: `visible`,
               },
               overlay: {
@@ -98,7 +105,7 @@ class DefaultLayout extends React.Component {
                 bottom: `unset`,
                 minHeight: `100%`,
                 minWidth: `100%`,
-                zIndex: 10,
+                zIndex: zIndices.modal,
                 overflowY: `auto`,
                 backgroundColor: `rgba(255, 255, 255, 0.95)`,
               },
@@ -108,29 +115,30 @@ class DefaultLayout extends React.Component {
           >
             <div
               css={{
-                backgroundColor: `#ffffff`,
-                borderRadius: presets.radius,
-                boxShadow: `0 0 90px -24px ${colors.gatsby}`,
+                backgroundColor: colors.white,
+                borderRadius: radii[2],
+                boxShadow: shadows.dialog,
                 position: `relative`,
               }}
             >
               <button
                 onClick={this.handleCloseModal}
                 css={{
-                  background: colors.ui.bright,
+                  background: colors.white,
                   border: 0,
-                  borderBottomLeftRadius: presets.radius,
-                  borderTopRightRadius: presets.radius,
-                  color: colors.gatsby,
+                  borderRadius: radii[6],
+                  color: colors.text.secondary,
                   cursor: `pointer`,
                   position: `absolute`,
                   left: `auto`,
-                  right: 0,
+                  right: space[7],
+                  top: space[8],
                   height: 40,
                   width: 40,
+                  fontSize: fontSizes[4],
                   "&:hover": {
-                    background: colors.gatsby,
-                    color: `#fff`,
+                    background: colors.ui.hover,
+                    color: colors.gatsby,
                   },
                 }}
               >
@@ -141,56 +149,27 @@ class DefaultLayout extends React.Component {
               {this.props.modalNextLink}
             </div>
           </Modal>
-        </React.Fragment>
+        </>
       )
     }
 
     return (
-      <div className={isHomepage ? `is-homepage` : ``}>
-        <Helmet defaultTitle={`GatsbyJS`} titleTemplate={`%s | GatsbyJS`}>
-          <meta name="twitter:site" content="@gatsbyjs" />
-          <meta name="og:type" content="website" />
-          <meta name="og:site_name" content="GatsbyJS" />
-          <link
-            rel="canonical"
-            href={`https://gatsbyjs.org${this.props.location.pathname}`}
-          />
-          <html lang="en" />
-        </Helmet>
-        <Banner background={isHomepage ? `#402060` : false}>
-          These are the docs for v2.
-          {` `}
-          <a
-            href="https://v1.gatsbyjs.org/"
-            css={{
-              color: `#fff`,
-            }}
-          >
-            View the v1 docs
-            <span
-              css={{
-                display: `none`,
-                [presets.Mobile]: {
-                  display: `inline`,
-                },
-              }}
-            >
-              {` `}
-              instead
-            </span>
-          </a>
-          .
-        </Banner>
+      <>
+        <SiteMetadata pathname={this.props.location.pathname} />
+        <SkipNavLink css={skipLink}>Skip to main content</SkipNavLink>
+        <Banner />
         <Navigation pathname={this.props.location.pathname} />
         <div
           className={`main-body`}
           css={{
-            paddingTop: presets.bannerHeight,
-            [presets.Tablet]: {
-              margin: `0 auto`,
-              paddingTop: isHomepage
-                ? presets.bannerHeight
-                : `calc(${presets.bannerHeight} + ${presets.headerHeight})`,
+            paddingLeft: `env(safe-area-inset-left)`,
+            paddingRight: `env(safe-area-inset-right)`,
+            paddingTop: sizes.bannerHeight,
+            // make room for the mobile navigation
+            paddingBottom: sizes.headerHeight,
+            [breakpointGutter]: {
+              paddingTop: `calc(${sizes.bannerHeight} + ${sizes.headerHeight})`,
+              paddingBottom: 0,
             },
           }}
         >
@@ -203,7 +182,7 @@ class DefaultLayout extends React.Component {
           />
         </div>
         <MobileNavigation />
-      </div>
+      </>
     )
   }
 }