+++ www/src/views/starter-library/lhs-filter.js
@@ -1,10 +1,7 @@
 import React from "react"
-// import FaAngleDown from "react-icons/lib/fa/angle-down"
-// import FaAngleUp from "react-icons/lib/fa/angle-up"
 import MdCheckboxBlank from "react-icons/lib/md/check-box-outline-blank"
 import MdCheckbox from "react-icons/lib/md/check-box"
 import { colors } from "../../utils/presets"
-import { scale } from "../../utils/typography"
 
 import Collapsible from "../shared/collapsible"
 import styles from "../shared/styles"
@@ -15,9 +12,10 @@ export default function LHSFilter({
   data,
   filters,
   setFilters,
+  fixed,
 }) {
   return (
-    <Collapsible heading={heading}>
+    <Collapsible heading={heading} fixed={fixed}>
       {data
         .sort(([a, anum], [b, bnum]) => {
           if (sortRecent) {
@@ -44,9 +42,8 @@ export default function LHSFilter({
             >
               <div
                 css={{
-                  color: filters.has(c) ? colors.gatsby : colors.ui.bright,
-                  ...scale(0),
-                  marginRight: 8,
+                  color: filters.has(c) ? colors.gatsby : colors.input.border,
+                  ...styles.filterCheckbox,
                 }}
               >
                 {filters.has(c) ? (
@@ -69,4 +66,8 @@ export default function LHSFilter({
         ))}
     </Collapsible>
   )
+}
+
+LHSFilter.defaultProps = {
+  fixed: false,
 }