+++ packages/gatsby-plugin-lodash/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.1.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-lodash/compare/gatsby-plugin-lodash@3.0.5...gatsby-plugin-lodash@3.1.0) (2019-06-20)
+
+**Note:** Version bump only for package gatsby-plugin-lodash
+
+## [3.0.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-lodash/compare/gatsby-plugin-lodash@3.0.4...gatsby-plugin-lodash@3.0.5) (2019-03-11)
+
+**Note:** Version bump only for package gatsby-plugin-lodash
+
+## [3.0.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-lodash/compare/gatsby-plugin-lodash@3.0.3...gatsby-plugin-lodash@3.0.4) (2019-02-01)
+
+**Note:** Version bump only for package gatsby-plugin-lodash
+
+<a name="3.0.3"></a>
+
+## [3.0.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-lodash/compare/gatsby-plugin-lodash@3.0.2...gatsby-plugin-lodash@3.0.3) (2018-11-29)
+
+**Note:** Version bump only for package gatsby-plugin-lodash
+
+<a name="3.0.2"></a>
+
+## [3.0.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-lodash/compare/gatsby-plugin-lodash@3.0.1...gatsby-plugin-lodash@3.0.2) (2018-10-29)
+
+**Note:** Version bump only for package gatsby-plugin-lodash
+
 <a name="3.0.1"></a>
 
 ## [3.0.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-lodash/compare/gatsby-plugin-lodash@3.0.1-rc.1...gatsby-plugin-lodash@3.0.1) (2018-09-17)